Finished the project component on work page and added firebase config
Finished the project UI component that will be used to present the different projects on the work page. The component is composed of several small components in order to achieve modularity and make it easier to maintain and shift if needed.
Added firebase dependency in the project, as firebase will be used as a database for this project. initilaised the Firebase config file. All connection variables are stored in a .env file