Fixed the scroll and modified the Project class
fixed the scroll and modified the Project class. The project class now also has a member for the repository url . Modified the ProjectItemComponent to takin a property repoUrl, which is of type string and represents the repository of the project url address. On click on the project item, the user will be redirected to the github repo of the project.