Partially implemented the filter logic
Implemented parts of the filter logic.
Set up a state variable, that stores the filter criteria for programming languages and technologies.
If filter criteria is set, only the projects that pass the criteria are displayed on screen. The logic is executing on the WorkPage, where an every method is applied to the filterLanguages, checking if they are included in the stack array of each project. The logic only executes if the filterLanguages array is not empty (length bigger than 0).