Implemented the privacy filter and repurposed the PrivateLabel
Implemented the logic for the project privacy filter. Now projects can be filtered based on privacy too.
The privacy filter is implemented in the filter menu in the WorkPageHeader component.
The filter uses a state variable of type number. The filtering logic is implemented in the WorkPage comonent, in a useEffect hook. The hook handles both the language filtering and the privacy filtering. Repurposed the PrivateLabel component, now it's named PrivacyLabel and displays both the "private" and the "public" labels.
Implemented styling changes.