Finalised the portfolio
Added the About me text on the About me page in the AboutMeTextComponent.
Set up a new feature - FetchLoader, which is a component that displays on the work page, while projects are being fetched from the database.
Added an email icon in the SoMeIconsHolder component. The icon also has a mail to and abbr tag, that displays my email.
Improved the filter button, now when a filter is active it's background color is different.